added resolve_graph_entities
